requirements.txt
streamlit
opencv-python-headless
pytesseract
fuzzywuzzy
python-Levenshtein
numpy
pillow